jit as a library and binary
